requirements.txt
requests
beautifulsoup4
openai
python-dotenv
Flask
pandas
pydantic
google-genai
flask-cors
openpyxl
python-dotenv